bootify & webify
Created build.boot and added a "boot serve" task which will serve the app defined in riffmuse.web on localhost:3000 and also start a dev REPL.

Keeping Leiningen project.clj for now -- it's currently still needed for the "lein bin" plugin, which creates an executable file which can be placed in $PATH so the 'riffmuse' program can be run on the command-line. Will eventually use Boot's scripting awesomeness instead -- can add a "-main" method to build.boot and then users should be able to rename build.boot -> riffmuse and copy it to their $PATH.

Right now riffmuse.web just serves the result of (riffmuse.core/-main "E blues"), but will make it more robust, with an input field and a shiny button.