fix version number not showing up (issue #2)
Unfortunately there doesn't seem to be an easy way to pass in the version number from build.boot (i.e. +version+) to the program itself, so that the program can display the version number. I'm not sure yet about the best way to handle this with Boot.

The problem is that build.boot can't read the version number from the jar metadata (at least, not yet) like versioneer does (only works with Leiningen, at the moment), and the project can't read the version number from build.boot when run directly as a jar.

I'm working around it for the time being by basically doing it both ways -- when executing build.boot as a script, the build.boot -main method binds riffmuse.core/*version* to +version+; when running the project as a jar, versioneer is used to find the version number and bind it to *version*.

This kinda gives me the heeby-jeebies, but oh well... ¯\_(ツ)_/¯